выделить логику карточки фильма в отдельный презентер
